🌟 Changed Order class to a a published Order struct within OrderClass
structs conatining Codable properties basically come with free Codable conformance unlike Classes with published properties that need custom initializer and encoder methods.
Making the struct Codable and nesting it in an ObservableObject class with a published instance of the struct saves us from writing the custom implementation to support Codable.
Basically, the whole class is not getting archived and unarchived now, the order property of the class conforms to Codable and that is the one being archived and unarchived.